requirements.txt
# ML Pipeline - Python Dependencies
# ==================================

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn==1.3.2
joblib>=1.3.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Logging (built-in, no install needed)

# Optional: For future enhancements
# mlflow>=2.8.0
# great-expectations>=0.17.0
